#1: Move PoseDetector and VideoProcessor to pose_markup folder.
